+++ tooling/ast_fuzzer/src/program/types.rs
@@ -1,0 +1,138 @@
+use std::collections::HashSet;
+
+use acir::FieldElement;
+use noirc_frontend::{
+    ast::IntegerBitSize, hir_def, monomorphization::ast::Type, shared::Signedness,
+};
+use strum::IntoEnumIterator as _;
+
+/// Calculate the depth of a type.
+///
+/// Leaf types have a depth of 0.
+pub(crate) fn type_depth(typ: &Type) -> usize {
+    match typ {
+        Type::Field | Type::Bool | Type::String(_) | Type::Unit | Type::Integer(_, _) => 0,
+        Type::Array(_, typ) => 1 + type_depth(typ),
+        Type::Tuple(types) => 1 + types.iter().map(type_depth).max().unwrap_or_default(),
+        _ => unreachable!("unexpected type: {typ}"),
+    }
+}
+
+/// Collect all the sub-types produced by a type.
+///
+/// It's like a _power set_ of the type.
+pub(crate) fn types_produced(typ: &Type) -> HashSet<Type> {
+    /// Recursively visit subtypes.
+    fn visit(acc: &mut HashSet<Type>, typ: &Type) {
+        if acc.contains(typ) {
+            return;
+        }
+
+        // Trivially produce self.
+        acc.insert(typ.clone());
+
+        match typ {
+            Type::Array(len, typ) => {
+                if *len > 0 {
+                    visit(acc, typ);
+                }
+                // Technically we could produce `[T; N]` from `[S; N]` if
+                // we can produce `T` from `S`, but let's ignore that;
+                // instead we will produce `[T; N]` from any source that can
+                // supply `T`, one of which would be the `[S; N]` itself.
+                // So if we have `let foo = [1u32, 2u32];` and we need `[u64; 2]`
+                // we might generate `[foo[1] as u64, 3u64]` instead of "mapping"
+                // over the entire foo. Same goes for tuples.
+            }
+            Type::Tuple(types) => {
+                for typ in types {
+                    visit(acc, typ);
+                }
+            }
+            Type::String(_) => {
+                // Maybe it could produce substrings, but it would be an overkill to enumerate.
+            }
+            Type::Field => {
+                // There are `try_to_*` methods, but let's consider only what is safe.
+                acc.insert(Type::Integer(Signedness::Unsigned, IntegerBitSize::HundredTwentyEight));
+            }
+            Type::Integer(sign, integer_bit_size) => {
+                // Casting up is safe.
+                for size in IntegerBitSize::iter()
+                    .filter(|size| size.bit_size() > integer_bit_size.bit_size())
+                {
+                    acc.insert(Type::Integer(*sign, size));
+                }
+                // There are `From<u*>` for Field
+                if !sign.is_signed() {
+                    acc.insert(Type::Field);
+                }
+            }
+            Type::Bool => {
+                // Maybe we can also cast to u1 or u8 etc?
+                acc.insert(Type::Field);
+            }
+            Type::Slice(typ) => {
+                visit(acc, typ);
+            }
+            Type::Reference(typ, _) => {
+                visit(acc, typ);
+            }
+            Type::Function(_, ret, _, _) => {
+                visit(acc, ret);
+            }
+            Type::Unit | Type::FmtString(_, _) => {}
+        }
+    }
+
+    let mut acc = HashSet::new();
+    visit(&mut acc, typ);
+    acc
+}
+
+pub(crate) fn to_hir_type(typ: &Type) -> hir_def::types::Type {
+    use hir_def::types::{Kind as HirKind, Type as HirType};
+
+    // Meet the expectations of `Type::evaluate_to_u32`.
+    let size_const = |size: u32| {
+        Box::new(HirType::Constant(
+            FieldElement::from(size),
+            HirKind::Numeric(Box::new(HirType::Integer(
+                Signedness::Unsigned,
+                IntegerBitSize::ThirtyTwo,
+            ))),
+        ))
+    };
+
+    match typ {
+        Type::Unit => HirType::Unit,
+        Type::Bool => HirType::Bool,
+        Type::Field => HirType::FieldElement,
+        Type::Integer(signedness, integer_bit_size) => {
+            HirType::Integer(*signedness, *integer_bit_size)
+        }
+        Type::String(size) => HirType::String(size_const(*size)),
+        Type::Array(size, typ) => HirType::Array(size_const(*size), Box::new(to_hir_type(typ))),
+        Type::Tuple(items) => HirType::Tuple(items.iter().map(to_hir_type).collect()),
+        Type::FmtString(_, _)
+        | Type::Slice(_)
+        | Type::Reference(_, _)
+        | Type::Function(_, _, _, _) => {
+            unreachable!("unexpected type converting to HIR: {}", typ)
+        }
+    }
+}
+
+/// Check if the type works with `UnaryOp::Minus`
+pub(crate) fn is_signed(typ: &Type) -> bool {
+    match typ {
+        Type::Field => true,
+        Type::Integer(signedness, _) => signedness.is_signed(),
+        _ => false,
+    }
+}
+
+/// Check if the type works with `UnaryOp::Not`
+pub(crate) fn is_bool(typ: &Type) -> bool {
+    matches!(typ, Type::Bool)
+}
